refactor: Phase 3 slice 3f — storage/file/shell/settings/watch-progress/auto-download routers (#84)
Extracts the six low-risk leaf IPC domains from the monolithic
`registerIpcHandlers()` in `src/main/index.ts` into per-domain
`src/main/ipc/*.ipc.ts` routers (refactor epic #84, Phase 3 — #102).

23 handlers moved into 6 new routers:
- storage.ipc.ts        — STORAGE_PICK_HOT_DIR/COLD_DIR, STORAGE_MOVE_TO_COLD/GET_USAGE/RUN_CLEANUP
- file.ipc.ts           — FILE_CHECK_EPISODES/OPEN/SHOW_IN_FOLDER/DELETE_EPISODE
- shell.ipc.ts          — SHELL_OPEN_EXTERNAL/_FILE
- settings.ipc.ts       — GET_SETTING, SET_SETTING
- watch-progress.ipc.ts — WATCH_PROGRESS_SAVE/GET/GET_ALL
- auto-download.ipc.ts  — 7 AUTO_DL_* handlers

The episode-file scan cache (`fileCheckCache` + `scanEpisodeFiles` /
`backgroundRescan`) stays in `index.ts` because `coldStorageService`
also feeds off it; `file.ipc.ts` / `storage.ipc.ts` reach it through
three new injected `AppDeps` members (`checkEpisodeFiles`,
`invalidateFileCacheByDirName`, `clearFileCache`). `FileCheckResult`
moved to `ipc/index.ts`.

`index.ts` shed ~215 lines and dropped the now-dead `dialog` import and
the four `autoDl*` aliases. `registerIpcHandlers()` is down to the 22
player/syncplay/debug handlers (Phase 3 slice 3g).

Behavior-identical; no IPC contract change. DESIGN.md updated.
