+++ DESIGN.md
@@ -43,6 +43,12 @@ Renderer (Vue)  --ipcRenderer.invoke-->  Preload (bridge)  --ipcMain.handle-->  
 | `src/main/ipc/shikimori.ipc.ts` | Shikimori router (Phase 3 slice 3d): all 16 `CHANNELS.SHIKIMORI_*` handlers — OAuth (`GET_AUTH_URL`/`EXCHANGE_CODE`/`LOGOUT`/`GET_USER`), rate read/write (`GET_RATE`/`UPDATE_RATE` with offline-queue fallback on network error), `GET_FRIENDS_RATES`/`GET_ANIME_RATES` (cache-first + background refresh)/`GET_ANIME_DETAILS`/`TRIGGER_DETAIL_PREFETCH`, sync surface (`GET_OFFLINE_QUEUE_LENGTH`/`GET_SYNC_STATUS`/`TRIGGER_SYNC`), `GET_FRIENDS_ACTIVITY`/`GET_CALENDAR`/`GET_RELATED`. Holds the `fetchAndCacheShikimoriRates`/`refreshShikimoriRatesInBackground` orchestration locally |
 | `src/main/services/shikimori-sync/index.ts` | `createShikimoriSyncService` — owns the Shikimori sync timer + offline-queue drain worker (`syncShikimoriQueue`, consolidate → drift-resolve → reconcile), the anime-details prefetcher (`prefetchShikimoriDetails`/`refreshShikimoriDetailsForMalId`), and the 5-minute calendar cache. All sync/prefetch/calendar module state (formerly free vars in `index.ts`) now lives in the service closure. Deps-injected (`store`, `broadcast` + 4 channel names) |
 | `src/main/ipc/skip-detector.ipc.ts` | Skip-detector router (Phase 3 slice 3e): `CHANNELS.SKIP_DETECTOR_{GET_DETECTIONS, GET_STATUS, CANCEL, CACHE_STATS, ANALYZE_SHOW, DETECT_STREAM, CANCEL_STREAM_DETECT, BACKFILL_ALL, QUEUE_STATUS}` + `DOWNLOAD_INJECT_CHAPTERS` (deferred from 3c — `fluent-ffmpeg` `-map_chapters` write over each downloaded MKV, with a fingerprint-cache key rewrite to survive the size/mtime change) |
+| `src/main/ipc/storage.ipc.ts` | Storage router (Phase 3 slice 3f): `CHANNELS.{STORAGE_PICK_HOT_DIR, STORAGE_PICK_COLD_DIR, STORAGE_MOVE_TO_COLD, STORAGE_GET_USAGE, STORAGE_RUN_CLEANUP}` — hot/cold directory pickers (persist to `store`, retarget `downloadManager`) and the `coldStorageService` move/usage/cleanup passes; clears the file-scan cache via `clearFileCache()` before a bulk move |
+| `src/main/ipc/file.ipc.ts` | File router (Phase 3 slice 3f): `CHANNELS.{FILE_CHECK_EPISODES, FILE_OPEN, FILE_SHOW_IN_FOLDER, FILE_DELETE_EPISODE}` — episode-file scan via the injected `checkEpisodeFiles`, open/reveal via `shell`, delete cascading to `coldStorageService` + the skip-detector caches |
+| `src/main/ipc/shell.ipc.ts` | Shell router (Phase 3 slice 3f): `CHANNELS.{SHELL_OPEN_EXTERNAL, SHELL_OPEN_EXTERNAL_FILE}` — `shell.openExternal` / `shell.openPath` wrappers |
+| `src/main/ipc/settings.ipc.ts` | Settings router (Phase 3 slice 3f): `CHANNELS.{GET_SETTING, SET_SETTING}` — generic `store` get/set, with `downloadDir` resolved through `coldStorageService` |
+| `src/main/ipc/watch-progress.ipc.ts` | Watch-progress router (Phase 3 slice 3f): `CHANNELS.{WATCH_PROGRESS_SAVE, WATCH_PROGRESS_GET, WATCH_PROGRESS_GET_ALL}` — per-`animeId:episodeInt` position/duration/watched tracking in `store.watchProgress` |
+| `src/main/ipc/auto-download.ipc.ts` | Auto-download router (Phase 3 slice 3f): `CHANNELS.AUTO_DL_{GET_SUBSCRIPTION, SET_SUBSCRIPTION, LIST_SUBSCRIPTIONS, TRIGGER, GET_STATUS, GET_ENABLED, SET_ENABLED}` — subscription CRUD + tick triggers, delegating to the `auto-downloader` module |
 | `src/main/smotret-api.ts` | Smotret-Anime API client: search, anime/episode details, embed, subtitles, token validation |
 | `src/main/download-manager.ts` | Download queue, concurrent downloads, progress, ffmpeg merge, scan-merge |
 | `src/main/shikimori.ts` | Shikimori API client: OAuth, token refresh, user rates, anime list |
